+++ src/main/java/janggi/domain/path/path_filter/LastPositionAllyPathFilter.java
@@ -6,7 +6,7 @@ import janggi.domain.piece.Piece;
 import java.util.List;
 import java.util.Set;
 
-public class NonStopAtAllyPathFilter implements PathFilter {
+public class LastPositionAllyPathFilter implements PathFilter {
 
     @Override
     public void filter(final Piece piece, final Set<Path> paths, final List<Piece> allyPieces, final List<Piece> enemyPieces) {
