+++ src/main/java/janggi/domain/path/path_filter/LastPositionSameTypePathFilter.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
-public class NonStopAtSameTypePathFilter implements PathFilter {
+public class LastPositionSameTypePathFilter implements PathFilter {
 
     @Override
     public void filter(final Piece piece, final Set<Path> paths, final List<Piece> allyPieces, final List<Piece> enemyPieces) {
