+++ src/main/java/janggi/domain/path/path_provider/CrossPathProvider.java
@@ -6,7 +6,7 @@ import janggi.domain.position.Position;
 import java.util.HashSet;
 import java.util.Set;
 
-public class CrossLinePathProvider implements PathProvider {
+public class CrossPathProvider implements PathProvider {
 
     @Override
     public Set<Path> get(final Position position) {
