+++ src/main/java/janggi/domain/path/path_provider/movement_path_provider/CrossOneStepPathProvider.java
@@ -1,11 +1,10 @@
 package janggi.domain.path.path_provider.movement_path_provider;
 
-import janggi.domain.path.path_provider.MovementPathProvider;
 import janggi.domain.position.Movement;
 
 import java.util.Set;
 
-public class CrossPathProvider extends MovementPathProvider {
+public class CrossOneStepPathProvider extends MovementPathProvider {
 
     private static final Set<Movement> MOVEMENTS = Set.of(
             Movement.UP,
