+++ src/main/java/janggi/domain/path/path_provider/movement_path_provider/StraightDiagonalDiagonalPathProvider.java
@@ -1,6 +1,5 @@
 package janggi.domain.path.path_provider.movement_path_provider;
 
-import janggi.domain.path.path_provider.MovementPathProvider;
 import janggi.domain.position.Movement;
 
 import java.util.Set;
@@ -8,14 +7,15 @@ import java.util.Set;
 public class StraightDiagonalDiagonalPathProvider extends MovementPathProvider {
 
     private static final Set<Movement> MOVEMENTS = Set.of(
+            Movement.UP_UPLEFT_UPLEFT,
+            Movement.UP_UPRIGHT_UPRIGHT,
             Movement.DOWN_DOWNLEFT_DOWNLEFT,
             Movement.DOWN_DOWNRIGHT_DOWNRIGHT,
-            Movement.LEFT_DOWNLEFT_DOWNLEFT,
             Movement.LEFT_UPLEFT_UPLEFT,
-            Movement.RIGHT_DOWNRIGHT_DOWNRIGHT,
+            Movement.LEFT_DOWNLEFT_DOWNLEFT,
             Movement.RIGHT_UPRIGHT_UPRIGHT,
-            Movement.UP_UPLEFT_UPLEFT,
-            Movement.UP_UPRIGHT_UPRIGHT
+            Movement.RIGHT_DOWNRIGHT_DOWNRIGHT
+
     );
 
     @Override
