+++ src/main/java/janggi/domain/path/path_provider/movement_path_provider/StraightDiagonalPathProvider.java
@@ -1,6 +1,5 @@
 package janggi.domain.path.path_provider.movement_path_provider;
 
-import janggi.domain.path.path_provider.MovementPathProvider;
 import janggi.domain.position.Movement;
 
 import java.util.Set;
@@ -10,12 +9,12 @@ public class StraightDiagonalPathProvider extends MovementPathProvider {
     private static final Set<Movement> MOVEMENTS = Set.of(
             Movement.UP_UPLEFT,
             Movement.UP_UPRIGHT,
+            Movement.DOWN_DOWNLEFT,
+            Movement.DOWN_DOWNRIGHT,
             Movement.LEFT_UPLEFT,
             Movement.LEFT_DOWNLEFT,
-            Movement.RIGHT_DOWNRIGHT,
             Movement.RIGHT_UPRIGHT,
-            Movement.DOWN_DOWNLEFT,
-            Movement.DOWN_DOWNRIGHT
+            Movement.RIGHT_DOWNRIGHT
     );
 
     @Override
